Fixed main page TOC link to tutorial on changing notes to rests.
Explanation: this was an annoying one to catch.

Here's the relevant part of the main page TOC:

    tutorials/changing_notes_to_rests/index
    tutorials/creating_rest_delimited_slurs/index
    tutorials/making_grob_overrides/index
    tutorials/understanding_lilypond_grobs/index
    tutorials/understanding_time_signature_marks/index
    tutorials/working_with_component_parentage/index
    tutorials/working_with_threads/index

The error was that the first entry (for changing note to rests)
was preceeded by a TAB instead of SPACES.

That was enough to cause Sphinx to be unable to recognize the entry.