Added a change log to the appendices section of the docs.
The change log externalizes the large number of API changes from 2.9 to 2.10.

This cleans up the version history presented near the top of the doc site.

Changes from 2.8 to 2.9 and between other versions will also
eventually migrate out of the version history doc and into the change log.

This will create a relatively clean separation of content whereby new functionality
is showcased in the version history document while changes are documented
in the chnage log.