Extended OffsetPositionedRhythmExpression.rotate() with a new 'fracture_spanners=True' keyword.
Set fracture_spanners=False to leave spanners alone during rotation.
This has the effect of allowing beams not to break during rotation.

This is the first time such leave-spanners-alone-during-rotation functionality
has appeared in the system. It seems surprising that such a thing is possible.
The implementation is based on mutating components in the score tree and then
sorting the components of all spanners according to the (new) score index
of every spanned component. This is a very genernal pattern. And it means
that it may be possible to implement other score tree mutation functions
that seemed difficult or impossible before now.