Redefined part of the rhythm library.
Libraryized rhythm makers previously accepted an optional 'beam' attribute.
The name of the attribute was cryptic. The meaning of the attribute was
to beam together all cells produced in a single call to the maker.

The new name of the attribute is 'beam_cells_together'.

This is part of a pattern that will allow for the following two properties on all rhythm makers:

    RhythmMaker.beam_each_cell      [True, False]
    RhythmMaker.beam_cells_together [True, False]