Extended tuplettools.make_tuplet_from_duration_and_proportions(duration, proportions) to coerce duration input.
Does not need to be noted in the 2.11 release notes because all functions everywhere in the API
that take a duration argument should their duration arguments.

In other words, we support a principle of universal (duration) coercion.

So any functions lacking the behavior should implement the behavior.

* * *

Also added LayoutBlock test to check for LilyPond command mark formatting.