Extended SegmentSpecification.set_rhythm() to acceptable aribtrarily parseable strings for rhythmic input.
Example:

    score_template = scoretemplatetools.GroupedRhythmicStavesScoreTemplate(staff_count=2)
    score_specification = specificationtools.ScoreSpecification(score_template)

    red_segment = score_specification.make_segment(name='red')

    red_segment.set_time_signatures([(2, 8), (2, 8), (3, 8)])
    red_segment_rhythm = "{ c'16 [ ( c'8 c'8. ] ) }"
    red_segment.set_rhythm(red_segment_rhythm, contexts=['Voice 1'])
    red_segment.set_rhythm(red_segment_rhythm, contexts=['Voice 2'], reverse=True)

    score = score_specification.interpret()

The string input is parsed with iotools.p() directly.

This has the incredible side effect of making beams and slurs (and so on) work for rhythmic input.

Just like with the rest of the specification system, input is repeated cyclically as
necessary to account for the necessary musical time of the specification.

Set-time reverse and rotation keywords are also available.