Added Multiplier class to the durationtools package.
The class is a literal alias of Duration.

Added Multiplier to the global namespace in abjad/__init__.py.

Class will be worked into functions throughout the public API as a replacement for built-in fraction.

A major advantage multiplier instances hold over fractions is that multipliers will coerce input.

(Built-in fractions do not coerce input.)