Made Component class abstract.
This happened in adding a 'duration_in_seconds' abstract property to Component.

Does not need to be noted in the change log because Component was already being
treated as abstract before this change. It was an oversight that no abstract
method or abstract property had previously been implemented on the class.