Deprecating durationtools.multiply_duration_pair().
Use NonreducedFraction.multiply_without_reducing() instead.

Function will be removed in a coming commit.

As of this commit the function now returns a (custom) nonreduced fraction rather than a (built-in) pair.

Note that this is one of the lines of change that's currently happening in durationtools:
the use of built-in tuples is being largely replaced by use of custom classes like
durations, multipliers and nonreduced fractions.