Fixed old bug in measuretools.replace_contents_of_measures_in_expr(expr, new_contents).
The function previous stripped ties from new_contents.

The function now preserves ties on new_contents.

This was accomplished by adding a new Container private method:

    Container._append_without_withdrawing_from_crossing_spanners()

The method is composer-unsafe (as noted in the classfile).

The method is called only by measuretools.replace_contents_of_measures_in_expr().

It is conceivable that Container._append_without_withdrawing_from_crossing_spanners()
will be useful in the implementation of new functions to be added to the API in future.

It is also conceivable that Container._append_without_withdrawing_from_crossing_spanners()
will be useful in the tightening of existing API functions.