Renamed timetools.SymbolicTimepoint.
The new name is timetools.SymbolicOffset.

This commit begins the process of renaming 'timepoint' to 'offset' everywhere in the experimental branch.

The goal of the work is to harmonize the naming (and implementation) of experimental timetools
such that the package can be merged into the mainline.

The timetools package will be a deployment level feature the 2.11 build.