Extended Tuplet with implied_multiplier property.
What's going on here are a couple of commits to regularize
the relationship between 'multiplier' properties
and 'implied_prolation' properties on different system objects.

The way it will work is that 'multiplier' properties will
be largely read / write. This means that the user-settable
part of the interface for these things will be available
through multiplier manipulation.

The implied_prolation property, OTOH, will always be read-only.

Objects that carry a multiplier property will always have a
(basically redundant) implied_prolation property.

OTOH objects with a (read-only) implied_prolation property 
may or may not carry a (user-settable) multiplier property.