Added a new parseable tag to abjad-book:
    * <abjadextract module \>[flags]

This single-line tag imports the code found at `module`, and copies the
actual code text itself into the abjad-book session, just as though it
had been manually included between a pair of <abjad></abjad> tags.

The intended use is in Abjad's literature examples.  Most of the examples
are also written up in the demos/ directory.  It would be best if all of
the code pertaining to each literature example was housed entirely in the
demos/ directory, along with tests to guarantee their validity.  Then, 
using the name <abjadextract \> tag, that code can simply be included into
the documentation narrative.

This will reduce code duplication (and code drift) between the demos and
examples.

NOTE: This does NOT need to be mentioned in the 2.11 release notes.