Streamlined interpreter.
Removed ConcreteInterpreter.reestablish_rhythm_material_request_offsets() method.

This is simplification that is part of a debugging process.

The debugging process is designed to allow the interpretation of rhythm material requests
in two different contexts.

The first context is that in which a rhythm material request is being made
against rhythmic material in the *past*. This is already implemented and tested.

The second (newer) context is that in which a rhythm material request is being allowed
to repeat over several consecutive segments. The current implementation of the
interpreter doesn't yet allow this.

This commit gets closer to affording the second case.