Retired the terms 'score start offset', 'score stop offset' and 'score offset'.
Use 'start offset', 'stop offset' and 'offset' instead.

Explanation: what was going on was ambiguity between two things.
Those two things are now modeled by the (old) idea of an *offset*
versus the (new) idea of a *symbolic offset*.

What's the difference between these two things?

An offset is always an Offset object. Offset objects are essentially rationals.

A symbolic offset is an instance of the SymbolicOffset class.
It requires an interpreter to interpet a symbolic offset and
turn it into a (numeric or literal) offset.

A good way to reflect on this difference is to not that the (old)
idea of a (numeric) offset is implemented all over the mainline
while the (new) idea of a symbolic offset exists only in the experimental branch.