Changed spanner format behavior for leaves.
Consider the following situation:

    >>> staff = Staff("c'1 c'1")
    >>> clef = contexttools.ClefMark('bass')(staff.leaves[0])
    >>> trill = spannertools.TrillSpanner(staff[:])
    >>> trill.written_pitch = "b'"

The desired output is:

    \new Staff {
        \clef "bass"
        \pitchedTrill
        c'1 \startTrillSpan b'
        c'1 \stopTrillSpan
    }

However, until this commit, the actual (and faulty) output would be:

    \new Staff {
        \pitchedTrill
        \clef "bass"
        c'1 \startTrillSpan b'
        c'1 \stopTrillSpan
    }

This is because spanner format contributions which occur before a leaf
would be placed in the "before" format slot, and context mark contributions
would be placed in the "opening" format slot.  When formatting leaves,
this causes the context mark contributions to appear closer to the leaf
nucleus than the spanner format contributions (and likewise in a mirror image
for the "closing" and "after" slots), as "before" contributions always appear
earlier than "opening" contributions.

In this commit, formattools.get_spanner_format_contributions() now checks
if the component it is formatting is a Leaf or a Container.  If the component
is a Leaf, it uses "opening" instead of "before" and "closing" instead of
"after" for its spanner contributions.

This then results in the desired, format LilyPond example.

All tests pass.  All doctests pass.