Added new EvenRunRhythmMaker class.
For each division on which the class is called, the class
produces an even run of notes each equal in duration to 1/d
with d equal to the division denominator.

Example:

    >>> maker = rhythmmakertools.EvenRunRhythmMaker()

    >>> divisions = [(4, 16), (3, 8), (2, 8)]
    >>> lists = maker(divisions)
    >>> containers = sequencetools.flatten_sequence(lists)

    >>> staff = Staff(containers)

    >>> f(staff)
    \new Staff {
        {
            c'16 [
            c'16
            c'16
            c'16 ]
        }
        {
            c'8 [
            c'8
            c'8 ]
        }
        {
            c'8 [
            c'8 ]
        }
    }

The EvenRunRhythmMaker accepts an optional nonnegative integer denominator_multiplier_exponent=0
defaulting to zero. This value causes the division returned to be filled with double or 
quadruple or 2**n more notes than when the value is not set.