Added abjad-book Sphinx extension to the experimental directory.
Currently, the extension is able to examine resolved doctrees if two criteria are met:

    * the `abjadbook_should_process` flag is set in Sphinx's conf.py
    * the docname of the doctree to process starts with the `abjadbook_transform_path` value,
      as set in Sphinx's conf.py

This allows us to turn abjad-book processing on and off for documentation, and to limit that
processing to a specific subdirectory or directories.

The extension traverses each valid doctree and extracts all literal_block elements into a list.
These are examined, and their prompts stripped, creating valid commandline input.
Finally, the stripped text is searched for any occurances of "show(", which should trigger abjad-book processing.

Actual processing by abjad-book is not yet implemented.