abjad-book Sphinx extension can now create PNGs, and insert them into Sphinx output.
Note that this is not yet ideal behavior - we've yet to solve the problem of versioning images.

It may be fruitful to move towards the sort of system that the Sphinx inheritance diagram extension uses,
whereby images are created solely in the build directory when creating output for some target format.

That eliminates all problems with versioning, but may also incur significant processing overhead, especially
given how frequently Abjad developers rebuild their documentation.

As before, see the HTML page for lyrictools.Lyrics for proof.