Renamed input parameter 'prolated_duration' to 'duration' in last remaining spots in codebase.
The use of 'prolated_duration' as an input parameter name is deprecated. Just use 'duration' instead.

The docstring to the relevant function can then indicate that 'duration' is treated as prolated duration.

Note that this brings up a larger pattern that the entire codebase might be simplified
by always treating unqualified 'duration' to mean prolated duration. Pretty much the only
places where preprolated durations are necessary are the places dealing with the written durations
of notes, rests and chords. So working with duration almost always means working with prolated duration.
And those few places that deal with the written durations of notes, rests and chords
can say so explicitly.

If we decide to acknowledge this observation then 'prolated_duration' could probably be
be replaced with just 'duration' in all functions in the API.