Extended SegmentSpecification.request_rhythm() with selection=None keyword.
When selection is set then the timespan of the explicit selection will be used.
When selection=None the timespan of the entire segment will be supplied.

This enambles SegmentSpecification.request_rhythm() to function
exactly like ScoreSpecification.requet_rhythm() when the composer desires.

The only difference between the methods is the default selection
supplied when selection=None.