Renamed segment specification and score specification request method 'selector' keyword.
The keyword is now named 'timespan' instead of 'selector'.

As central as the notion of the selector is to the specificationtools system,
the label 'selector' is now mostly deprecated. Why? Because the label 'selector'
does nothing to indicate that the primary purpose of the object is to
select *horizontal* amounts of time (completely independently of vertical voices).