Added stub 'time_relation' property to MaterialRequest.
Added stub time_relation=None keyword to specification request methods.

The property and corresponding keywords are inert as of this commit.

Future commits will activate the property and corresponding keywords.

This will be done by deprecating time_relation property currently implement
on the selectors passed in to the specification request methods' timespan=None.

The larger goal here is a destruction of the concept of the Selector.
Why? Because the concept of the selector is currently a bundled (timespan, time_relation) pair.
Bundling these two things together previously seemed like a good idea. 
But it is now clear that it provides greater composer flexibility and a more
literate object model to unbundle timespans and time relations everywhere they occur in the system.
The primary place where timespans and time relations are currently bundled
is in the concept of the Selector. So the following commits will work
to undo the notion of the select and thereby unbundle timespans from the time relations
with which they interact.