Added 'start_segment_name' property to Setting. Added 'start_segment_name' property to SingleContxtSetting. Addeed 'start_segment_name' property to MultipleContextSetting.
The properties may be only temporary.

What's going on right now is the unbundling of timespan and time relation.
Part of doing this means that the parts of the system that extract "start segment"
from the various selector instances passed around the system may no longer
rely on selector instances containing start segment information.
Instead, more objects in the system will have to carry around "start segment"
information as a first-class piece of information.

It may in the long term be possible to remove all this "start segment"
information altogether. But the idea is currently deeply embedded and
will be preserved during the unbundling of timespan and time relation.