BackgroundMeasureSymbolicTimespan now implements an anchor=None keyword.
All TimeRelationSymbolicTimespan classes will eventually also implement this same anchor=None keyword.

The keyword is being introduced as part of the work of unbundling timespan and time relation information
in the system.

TimeRelationSymbolicTimespan classes currently store a 'time_relation' property.
This time relation is a TimespanTimespanTimeRelation that has been partially "loaded".
The partial loading refers to the presence of a "timespan_1" element already loaded in the time relation.
The work of unbundling now going on will seek to store this 'time_relation' property
as a completely unloaded time relation. What is now being stored as "timespan_1" will
be moved to the "anchor" property newly introduced in this commit.