Added new documentation tool:
    * documentationtools.write_expr_to_md5_hashed_ly()

Writes a score expression into a .ly file whose name is the MD5 hash of the LilyPond format of that expression.

This will eventually be used by the abjad-book Sphinx extension to detect "repeat" notation throughout the API, in order to save CPU on repeat renderings of identical scores (like in inherited property docstrings), and to save harddisk space overall.  The inheritance-diagrams extension works following a similar principle of hashing the graphviz .DOT descriptions of the diagrams, and using those hashes to enforce uniqueness.

NOTE: This does *not* have to be mentioned in the 2.11 release notes.