Moved select methods up in the inheritance hierarchy.
There are four select methods:

    self.select_background_measures()
    self.select_divisions()
    self.select_leaves()
    self.select_notes_and_chords()

These were previously implemented on Specification.
These are now implemented on SymbolicTimespan.

What does this mean?

This means something conceptually very important in terms of the system object model.
The point has to do with selecting things. The concept of what it means
to select stuff is one of the two or three most important things being
researched as part of the work of building specificationtools.
The move of the four select methods from Specification up to SymbolicTimespan
is an acknowledgment of the fact that a selection is always *made against something*.
And the something against which a selection is made always turns out to be a symbolic timespan.
It is because of this fact that it is possible to make object-oriented selections
against the SymbolicTimespan class.