Implemented the idea of a 'moniker' on symbolic timespans.
Symbolic timespans are frequently written to disk or otherwise held 
in a frozen form prior to evaluation.

Vanilla symbolic timespans can be frozen using their storage format.
But score and segment specifications are frozen as none and as a string, respectively.
The idea of the moniker generalizes all of this and provides just
the right amount of information for a symbolic timespan to be frozen.