Changed Setting.selector to Setting.anchor. IMPORTANT object model notes follow.
EXPLANATION: The object model now distinguishes clearly between these three concepts:

    1. symbolic timespan
    2. selector
    3. anchor

DETAILS (worth reading):

1. SYMBOLIC TIMESPAN. A symbolic timespan is (finally) exactly what you'd expect:
an object that models an expression that can be evaluated to produce a (literal)
timespan (featuring a start offset and a stop offset). How does a symbolic timespan
get evaluated? A symbolic timespan gets evaluated in the context of a score specification.
That's why there are a handful of methods in the system that pass around a score
specification as input parameter: these methods are essentially currying 
the evaluation contet of a symbolic timespan.

2. (OBJECT) SELECTOR. An (object) selector is a special device that lets composers
select background measures, divisions, notes, rests, chords, tuplets and so.
This sort of selection is *always done relative to a symbolic timespan*
and is always done *in the context of exactly one voice in score*.
Selectors can be evaluated when  given an appropriate score specification 
to serve as evaluation context. The result of evaluating a selector is
a list of zero or more background measures, divisions, notes, rests, chords,
tuplets or so on. Selectors are modeled as a type of symbolic timespan
because every selector evaluates to a list of objects that implies a timespan
(in exactly the same way that every selectiontools.Selection object implies a timespan).

3. ANCHOR. There is no anchor class in the system. The reason for this is that the
concept of an anchor does not refer to a class of object. The concept of an
anchor instead refers to *a way of using* a symbolic timespan. That is,
any symbolic timespan can be used as the anchor of any of other symbolic timespan.
Every selector anchors to some other symbolic timepan, for example.
So the concept of the anchor is very much a *syntactic* concept rather
than a concept of *essence*. Anchoring, that is, talks about *the way something is used*
rather than *what something is*. And the reason for this is that all anchors
are definitionally symbolic timespans.