Implemented a read-only 'modifications' list on all symbolic timespans.
The function of the list will be to hold an (arbitrarily long) list
of callback-like modifications to be applied to a symbolic timespan
during evaluation.

What might this look like?

This might mean moving the start offset or stop offset of the timespan
(ie, shortening or lengthing the timespan or translating the timespan in time).
It might mean scaling the timespan.

So the modifications list should obviate the need for the "operator" classes
now housed in the symbolictimetools package.

The modifications list should also eventually completely replace
the index=None, count=None, reverse=None, rotation=None, callback=None
keywords scattered all over the system. The new and improved
way of doing such things will be to simply load as many such
modifications as desired into the modifications list at specification time.