Added contexts scrape to `abjad/ly` as `abjad/ly/ly/contexts.ly` and `abjad/ly/contexts.py`.
The contexts scrape includes not only the names of all default contexts in LilyPond, but also what contexts each context accepts, each context's aliases, what engravers they consist of, and what grobs they can create.

With this last addition, we now have information on the vast bulk of LilyPond's built in environment.

The various Python files in abjad/ly allow us to do things like look up whether a Grob can be created in a particular context, then check for what interfaces the Grob employs, and _then_ look up a particular interface and find the properties it contains.

This means that we can finally ditch the hard-coded lists of context names and grob properties that are scattered throughout lilypondproxytools, and we can employ more robust checking of names and arguments if we so desire.  Given some of the LilyPond syntax changes slowly coming together, it may make sense to insist on a minimum LilyPond version, and then guarantee that our environment scrape mirrors that.