Improved timesignaturetools.establish_metrical_hierarchy() docs.
Also discovered a new subtlety about Sphinx:

If you attempt to write an ReST python code block, and indent your code, but neglect to preface it with ::, Sphinx will still format the code as though it was a Python code block.

However, the abjad-book Sphinx extension will not recognize the code block as a code block, and will ignore it.  Any changes you believe you've made to abjad-book's Python interpreting environment will be ignored.

This manifested for me by seeing the exact same notation displayed for three show() statements in a row.  The reason, as hinted at above, was because the intervening code blocks - where I redefined or otherwise altered the score objects I wanted to display - were not run in abjad-book's interpreting environment at all.

Proofread!