Removed selector=None keyword from system in favor of timespan=None keyword.
Explanation: according to the recent clean-up of the object model, selectors
are just a special type of (symbolic) timespan. This means that methods
that allow for optional filtering on horizontal amounts of time all now
filter based on the (more broadly defined) notion of timespan rather than
on the (more narrowly defined) notion of selector.