Moved set methods up from SegmentSpecification to SymbolicTimespan.
The motivation here is powerful. It will soon be possible to set things on *any* (symbolic) timespan.

This will soon work:

    measures = red_segment.select_background_measures(5, 10)
    measures.set_rhythm(library.sixteenths)

Notice that what's going on here is setting (a musical material) directly
on an arbitrary (symbolic) timespan. This is yet another large increase
in the overall object orientation of the system and will be enormously powerful.