Added new io tool:
    * iotools.graph()

A small number of classes throughout the system have started to gain a `graphviz_format` attribute, including datastructuretools.Digraph, documentationtools.InheritanceGraph, some of the rhythmtreetools.RhythmTreeNode subclasses, and even timesignaturetools.MetricalHierarchy.

Try it out:

::

    >>> hierarchy = timesignaturetools.MetricalHierarchy((7, 4))
    >>> iotools.graph(hierarchy)

It makes sense that we should be able to see these graph realizations just as easily as we see notation (or hear MIDI).

Furthermore, anyone successfully building our docs already has the graphviz toolchain installed in order to see the inheritance diagrams in the API.

With any luck, this can be stitched into the abjad-book functionality.