Renamed SymbolicTimespan._moniker to SymbolicTimespan._timespan_abbreviation.
The new name is less cryptic.

The purpose of the read-only private property is to turn a (symbolic) timespan
into format suitable for functioning as the anchor of other (symbolic) timespans.

What this comes out to mean is that only ScoreSpecification and SegmentSpecification
have an abbreviated timespan form (the timespan abbreviation of any ScoreSpecification is none
while the timespan abbreviation of any SegmentSpecification is the name of the segment).
The timespan abbreviation of all other (symbolic) timespans is simply the (symbolic)
timepsan itself.