Moved demos to same section of experimental API as the other sections.
The demos directory isn't really the same thing as a tools package.
So maybe the demos directory should be suppressed from showing up in the
experimental API?

But it's nice to be able to see the demo in the form that it currently
exists in. This helps make clear what work or additions are still required.
So I think it makes sense for the incomplete demos in the experimental
line to be visible somewhere in the experimental docs.