Added stub BeatSelector class. Modeling note included.
The class will allow the (thinly implemented) functionality of request_naive_beats()
to migrate to a selector.

MODEL: what is meant by 'naive' beats is a simple decomposition of time signature n/d
into n even "beats" of 1/d. So 6/8 return six "beats" of 1/8 duration each.
This doesn't have anything to do with any realworld understanding of meter. 
The functionality should instead be understood as a foothold for the eventual
integration of the full complement of realworld timing objects afforded by MetricalHierarchy.