Extended Timespan.set_offsets() to handle negative offset values. Model update included.
Explanation:

Timespan.set_offsets() *sets* positive values of start_offset and stop_offset:

    >>> timespan = timespantools.Timespan((1, 2), (3, 2)) 

    >>> timespan    
    Timespan(start_offset=Offset(1, 2), stop_offset=Offset(3, 2))

    >>> timespan.set_offsets(stop_offset=Offset(7, 8))
    Timespan(start_offset=Offset(1, 2), stop_offset=Offset(7, 8))

Timespan.set_offsets() *subtracts* negative values of start_offset and stop_offset:

    >>> timespan = timespantools.Timespan((1, 2), (3, 2)) 

    >>> timespan    
    Timespan(start_offset=Offset(1, 2), stop_offset=Offset(3, 2))

    >>> timespan.set_offsets(stop_offset=Offset(-1, 2))
    Timespan(start_offset=Offset(1, 2), stop_offset=Offset(1, 1))

MODEL: the analogy here is to the way that built-in slicing handles negative input arguments.
That is, slicing interprets positive arguments from left of an iterable and negative
arguments from the right of an iterable.