Rewired Selector to inherit from both SymbolicTimespan and Request. Model update included.
MODEL: the important change in this commit is that selectors now *are* requests.
This makes sense intuitively because both selectors and requests model the idea
of going out and getting discrete objects. How do selectors and requests differ?
Because a selector is a symbolic timespan a selector goes out and gets
discrete objects *relative to an Abjad score*. Because requests are not
symbolic timespans requests do not go get discrete objects relative to an Abjad score.
Rather, requests go get discrete objects that are not necessarily relative
to an Abjad score. The conrete request classes are AbsoluteRequest,
CommandRequest, HandlerRequest and StatalServerRequest. The request classes
fetch discrete pieces of information for nonscore sources.

Important to note here is that the unification of selectors and requests
means that there is now a notion of a 'modifications interface':

    Request.__getitem__()
    Request.partition_objects_by_ratio()
    Request.partition_objects_by_ratio_of_durations()
    Request.repeat_to_length()
    Request.reverse()
    Request.rotate()