Extended Timespan with is_tangent_to() predicate. Model update.
Example:

    >>> timespan_1 = timespantools.Timespan(5, 10)
    >>> timespan_2 = timespantools.Timespan(10, 15)

    >>> timespan_1.is_tangent_to(timespan_2)
    True

    >>> timespan_2.is_tangent_to(timespan_1)
    True    

    >>> timespan_1.is_tangent_to('text')
    False

MODEL: Because the predicate is bound to Timespan this begs
the question of whether something like all of the predicates
in timerelationtools should also be bound to Timespan.
I think the answer is probably yes. But even if the answer is no,
the new realization is that the (couple of dozen) functions in
timerelationtools constitute something like a "timespan inequality interface".

QUESTION: do we have existing language somewhere in the system (perhaps
timeitervaltools?) that distinguishes between "tangent to the left"
and "tangent to the right" of a timespan? (Because the tangent predicate
implemented here is reflexive.)