Renamed Request._slice_selected_objects() to Request.___getitem__().
Note that the extra initial underscore is added to the evaluable
implementations of delayed evaluation functions.

So selector.reverse() results in _reverse() being stored in the selector's modifications stack.
and selector.rotate() results in _rotate() being store in the selector's modifications stack.

The rename in this commit causes a call to __getitem__() to store ___getitem__()
with the extra underscore in the selector's modifications stack.