Removed VoiceSelector class. Concrete selectors all now inherit from Selector directly.
MODEL: all selectors are now defined against a voice.
To understand what this means, first consider that a later version of the system will implement 
"voice intermittency". What will voice intermittency be? Voice intermittency will
be the (possibly unexpected) appearance and subsequent disappearance of a voice at
arbitrary points in segment (or in the entire score). Note that voice intermittency
isn't anything particulary fancy notationwise: just usual case of secondary voices
popping up for only a couple of notes all over the place in the piano music of Brahms, for example.
The relevant point at present is that the system doesn't yet implement voice intermittency.
But when it does come time to implement voice intermittency the concept will be implemented
in such a way that *all* voices always "exist" for the entire duration of the score,
just as is assumed by the current implementation of the system. It helps to reflect
on these considerations to understand why the current commit attaches a voice name
to all selectors. A selector will always be taken relative to a voice. And every voice
will always be taken to run for the full duration of the score (even when voices
are, much later, allowed to appear and disappear intermittently in the final score).