Renamed symbolictimetools. The new name is timeexpressiontools.
MODEL: the new name better reflects what the classes in the package model:
arbitrary expressions that can be evaluated to equal either a (literal) timespan
or else a (literal) offset.

MODEL: the name also helps to achieve a modeling goal of using 'timespan'
to refer as nearly as possible to only Timespan objects. We now have a very
powerful model of a what a timespan is. The clarity of the model is muddied
if there are (qualified) types of timespan like 'symbolic' timespan.
(And a future commit will change SymbolicTimespan to TimespanExpression.
The word 'Timespan' will be present. But the rightmost word in the noun pile
will then be 'Expression'. And this matters because the hundreds of classes
in the system are all named in such a way that the most important word
in the noun pile is always at the extreme right. In other words, it should
be completely clear that a TimespanExpression is not a timespan; rather,
a TimespanExpression is an expression; what type of expression?;
and expression that evaluates to a timespan, which has been the point
of the class since introduction.)