Reimplemented documentationtools.InheritanceGraph's class organizing logic.
All considered classes are now organized into two dictionaries, with each class as a key, and a set of 0 or more classes as the value:

    * parent : children
    * child : parents

If any root classes were specified via `root_addresses`, only those classes which inherit from (or include) any of the root classes are included in the mappings.

If any lineage classes were specified via `lineage_addresses`, only those classes which either inherit from any lineage class, or from whom any lineage class inherits are included.

Root classes and lineage classes may be specified together or separately.

This makes it possible to visualize a "vertical slice" through the entire class hierarchy of Abjad as it passes through the classes defined in any given tools module.

The new graphviz_format for InheritanceGraph is yet to be implemented, and is still non-operational as of this commit.