Impelmented Timespan.fuse() and TimespanInventory.fuse() methods. Modeling commentary included.
Example with timespan:

    >>> timespan_1 = timespantools.Timespan(0, 5)
    >>> timespan_2 = timespantools.Timespan(5, 10)

    >>> timespan_1.fuse(timespan_2)
    Timespan(start_offset=Offset(0, 1), stop_offset=Offset(10, 1))

Example with timespan inventory:
    
    >>> example_inventory = timespantools.TimespanInventory()

    >>> example_inventory.append(timespantools.Timespan(0, 3))
    >>> example_inventory.append(timespantools.Timespan(3, 6))
    >>> example_inventory.append(timespantools.Timespan(6, 10))

    >>> example_inventory.fuse()

    >>> z(example_inventory)
    timespantools.TimespanInventory([
        timespantools.Timespan(
            start_offset=durationtools.Offset(0, 1),
            stop_offset=durationtools.Offset(10, 1)
            )
        ])

MODEL: these methods are similar to timeintervaltools.fuse_overlapping_intervals()
and timeintervaltools.fuse_tangent_or_overlapping_intervals(). The difference is
that Timespan.fuse() and TimespanInventory.fuse() insist on contiguous input.
So the corresponding timeintervaltools function would be timeintervaltools.fuse_tangent_intervals().

MODEL: the implementation of Timespan.fuse() means that all timespan-like objects
in the system (and especially those that inherit from Timespan) should consider
implementing a appropriate fuse() method.

MODEL: the differnet fuse variations available in the system beg the question
of what should be the result of taking the set-theoretic union of arbitrary timespans.