Specification now aggregates Timespan rather than inheriting from Timespan.
MODEL: Throughout the system we will make decisions as to whether any given
custom class should *be* a timespan or *have* a timespan. Components and spanners
have timespans; but specificatiotools-style settings are timespans.
The change committed here changes the ScoreSpecification and SegmentSpecification
from being timespans to merely having timespans. This gets at the heart of 
what the specification classes (ScoreSpecification and SegmentSpecification)
are supposed to be. The classes are ultimately supposed to be composer interface
classes: classes that give all the composition-time functionality composers will need.
Further revisions will probably recast ScoreSpecification and SegmentSpecification
as ScoreProxy and SegmentProxy, respectively, to separate out the remaining
different types of functionality implemented in each class.