Removed all references to 'symbolic' things.
Some objects are inherently symbolic (like time expressions). Others aren't.
This means that it is never necessary for a single class to implement
both 'symbolic_start_offset' (which would return an OffsetExpression)
and also (literal) 'start_offset' (which would return an Offset).
That is, many classes implement a 'start_offset' property;
some of those classes return a (literal) Offset for that property
and other classes return a (symbolic) OffsetExpression for that property;
and no classes return both; the distinction between the two
is maintained absolutely by the division of code into classes.