Doc update for previous commit.
All mainline documentation directories now contain their own index.rst file, acting as a TOC for that area.  This is very similar to the layout of the official Python docs, and makes it simpler to generate individual LaTeX PDFs for a single doc area if need be.

Also added the documentationtools.ReSTLineageDirective class.