Lineage diagrams are now constrained to at most 780 pixels wide in HTML output.
This makes some of them completely unreadable, but it's better than the previous layout. How might this be fixed? Diagrams could link to full-scale versions.  Still, the diagram for AbjadObject is shrunk so much as to be only a few pixels tall - it's not clear where many users could even click.

They could also perhaps "prune" the subclass subtrees somehow - children being the main source of extreme image dimensions, especially in classes like AbjadObject or Mark.